+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/viewholders/ParentMessageInfoViewHolder.kt
@@ -4,6 +4,7 @@ import android.view.View
 import com.sendbird.android.channel.BaseChannel
 import com.sendbird.android.channel.GroupChannel
 import com.sendbird.android.message.BaseMessage
+import com.sendbird.android.message.Emoji
 import com.sendbird.android.message.Reaction
 import com.sendbird.android.user.User
 import com.sendbird.uikit.activities.viewholder.GroupChannelMessageViewHolder
@@ -26,11 +27,13 @@ internal class ParentMessageInfoViewHolder(val binding: SbViewParentMessageInfoH
         emojiReactionLongClickListener: OnItemLongClickListener<String>?,
         moreButtonClickListener: View.OnClickListener?
     ) {
+        // not-used anymore
+    }
+
+    override fun setEmojiReaction(reactionList: List<Reaction>, totalEmojiList: List<Emoji>, emojiReactionClickListener: OnItemClickListener<String>?, emojiReactionLongClickListener: OnItemLongClickListener<String>?, moreButtonClickListener: View.OnClickListener?) {
         binding.parentMessageInfoView.binding.rvEmojiReactionList.apply {
-            setReactionList(reactionList)
-            setEmojiReactionClickListener(emojiReactionClickListener)
-            setEmojiReactionLongClickListener(emojiReactionLongClickListener)
-            setMoreButtonClickListener(moreButtonClickListener)
+            setReactionList(reactionList, totalEmojiList)
+            setClickListeners(emojiReactionClickListener, emojiReactionLongClickListener, moreButtonClickListener)
         }
     }